requirements.txt
Django==4.2.16
djangorestframework==3.15.2 
django-phonenumber-field==8.0.0 
phonenumbers==8.13.47 
psycopg2-binary==2.9.9 
Pillow==10.4.0 
django-filter==24.3
django-environ==0.11.2 
python-decouple==3.8
drf-yasg
djangorestframework-simplejwt
channels-redis
pytesseract
face_recognition
numpy
scikit-learn
scipy
joblib
jupyter
django-cors-headers
djangorestframework-simplejwt
django-filter